+++ src/utils/fieldWhenClause.ts
@@ -22,7 +22,10 @@ export const fieldWhenClause = (field: Field, parent: IMetadata, allFields?: Fie
     }
   }
 
-  const whenValue = parent[when.fieldRef];
+  let whenValue = parent[when.fieldRef];
+  if (whenValue === undefined) {
+    whenValue = field.default as string | IMetadata | string[] | null;
+  }
   if (when.caseSensitive || typeof when.caseSensitive === 'undefined') {
     return caseSensitive(when, field, whenValue);
   } else {
@@ -61,7 +64,7 @@ const caseInsensitive = (
  */
 const caseSensitive = (
   when: WhenClause,
-  field: Field,
+  _: Field,
   whenValue: string | IMetadata | string[] | null
 ) => {
   switch (when.operator) {